Separate baby DNA and pregnancy duration to allow burst compilation
