+++ Assets/Scripts/ECS/Reproduction/Pregnant.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a2bbabafee094ba40a095e25f0b3813b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
